Implementat clasa rezolvari cerinte + 2 cerinte query

+++ src/user/userNonPremium.cpp
@@ -14,7 +14,7 @@ UserNonPremium::UserNonPremium(){}
     costTransport = cost;
   }
 
-  int UserNonPremium::getCostTransport()
+  float UserNonPremium::getCostTransport()
   {
     return this->costTransport;
   }
